requirements.txt
Django==4.2
Pillow==9.5.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-allauth==0.58.1
django-tables2==2.7.0
django-filter==23.5
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
